+++ okitweb/static/okit/view/designer/js/artefacts/compartment.js
@@ -88,7 +88,7 @@ class CompartmentView extends OkitContainerDesignerArtefactView {
         return [DynamicRoutingGateway.getArtifactReference(), AutonomousDatabase.getArtifactReference(),
             ObjectStorageBucket.getArtifactReference(), FastConnect.getArtifactReference(),
             IpsecConnection.getArtifactReference(), RemotePeeringConnection.getArtifactReference(),
-            Drg.getArtifactReference()];
+            Drg.getArtifactReference(), NosqlDatabase.getArtifactReference()];
     }
 
     getRightEdgeArtifacts() {