+++ CHANGELOG.md
@@ -1,6 +1,32 @@
 # Release Notes
 
 
+## Version 0.39.0
+**Release Date**: 27th July 2022
+### Features
+1. Extend PCA Resource Query to include:
+    1. Groups
+    2. Policies
+    3. Users
+2. Add __namespace__ field to config to allow specification of the Object Storage namespace for PCA because the client.get_namespace is not implemented.
+3. Extend Tablular View to display additional information
+    1. Instance
+        1. Shape
+        2. Memory
+        3. OCPUs
+    2. Groups
+        1. Users
+        2. User Count
+4. New Properties sheet.
+    1. Analytics Instance
+5. New Resource
+    1. NoSQL Table / Index __(Creation only)__
+6. Remove Python module version specification and allow use of latest.
+### Bug Fixes
+1. Fix issue where PCA queried Compartment was not displayed as the top level compartment
+2. Resolve issues where documentation information was not be reloaded for resources.
+
+
 ## Version 0.38.0
 **Release Date**: 6th July 2022
 ### Features